dev: move CHI templates normalizer to appropriate module
