+++ pkg/model/chi/normalizer/templates/templater.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package normalizer
+package templates
 
 import (
 	log "github.com/altinity/clickhouse-operator/pkg/announcer"
